requirements.txt
Django==3.2.2
requests==2.25.1
django-filter==2.4.0
djangorestframework==3.12.4
django-cors-headers==3.14.0
# Swagger:
drf-yasg==1.21.0